feature: restructure the repository
Issue: #4

1. Created `backend` directory
2. Created `backend/api-service` and `backend/ai-service`
3. Updated the go.mod file
4. Run `go mod tidy`
